Calculates the Arrhenius-type temperature response

Given a kinetic rate at a reference temperature (argument tkref) this function calculates its temperature-scaling factor following Arrhenius kinetics.

calc_ftemp_arrh(tk, dha, tkref = 298.15)

Arguments

tk
Temperature (Kelvin)

dha
Activation energy (J mol-1)

tkref
Reference temperature (Kelvin)

Value

A numeric value for \(f\)

Details

To correct for effects by temperature following Arrhenius kinetics, and given a reference temperature \(T0\), \(f\) calculates the temperature scaling. Arrhenius kinetics are described by an equation of form \(x(T)= exp(c - \Delta Ha / (T R))\). The temperature-correction function \(f(T, \Delta Ha)\) is thus given by \(f=x(T)/x(T0)\) which is: $$ f = exp( \Delta Ha (T - T0) / (T0 R T_K) ) $$ \(\Delta Ha\) is given by argument dha. \(T\) is given by argument tk and has to be provided in Kelvin. \(R\) is the universal gas constant and is 8.3145 J mol-1 K-1. Note that this is equivalent to $$ f = exp( (\Delta Ha/R) (1/T0 - 1/T) ) $$

Examples

print( paste( "Relative rate change from 25 to 10 degrees Celsius (percent change):", (1.0-calc_ftemp_arrh( 283.15, 100000, tkref = 298.15))*100 ))

#> [1] "Relative rate change from 25 to 10 degrees Celsius (percent change): 88.1991209313034"