Calculates the temperature response of dark respiration

Given the dark respiration at the reference temperature 25 degress Celsius, this function calculates its temperature-scaling factor following Heskel et al. 2016.

calc_ftemp_inst_rd(tc)

Arguments

tc
Temperature (degrees Celsius)

Value

A numeric value for \(fr\)

Details

To correct for effects by temperature Heskel et al. 2016, and given the reference temperature \(T0 =\) 25 deg C, this calculates the temperature scaling factor to calculate dark respiration at temperature \(T\) (argument tc) as: $$ fr = exp( 0.1012 (T0-T) - 0.0005(T0^2 - T^2) ) $$ where \(T\) is given in degrees Celsius.

References

Heskel, M., O’Sullivan, O., Reich, P., Tjoelker, M., Weerasinghe, L., Penillard, A.,Egerton, J., Creek, D., Bloomfield, K., Xiang, J., Sinca, F., Stangl, Z., Martinez-De La Torre, A., Griffin, K., Huntingford, C., Hurry, V., Meir, P., Turnbull, M.,and Atkin, O.: Convergence in the temperature response of leaf respiration acrossbiomes and plant functional types, Proceedings of the National Academy of Sciences, 113, 3832–3837, doi:10.1073/pnas.1520282113,2016.

Examples

print(paste("Relative change in Rd going (instantaneously, i.e. not acclimatedly) from 10 to 25 degrees (percent change):", (calc_ftemp_inst_rd(25)/calc_ftemp_inst_rd(10)-1)*100 ))

#> [1] "Relative change in Rd going (instantaneously, i.e. not acclimatedly) from 10 to 25 degrees (percent change): 250.959273208264"