Calculates the temperature dependence of the quantum yield efficiency

Calculates the temperature dependence of the quantum yield efficiency following the temperature dependence of the maximum quantum yield of photosystem II in light-adapted tobacco leaves, determined by Bernacchi et al. (2003)

calc_ftemp_kphio(tc, c4 = FALSE)

Arguments

tc
Temperature, relevant for photosynthesis (degrees Celsius)

c4
Boolean specifying whether fitted temperature response for C4 plants is used. Defaults to FALSE (C3 photoynthesis temperature resposne following Bernacchi et al., 2003 is used).

Value

A numeric value for \(\phi(T)\)

Details

The temperature factor for C3 photosynthesis (argument c4 = FALSE) is calculated based on Bernacchi et al. (2003) as $$ \phi(T) = 0.352 + 0.022 T - 0.00034 T^2 $$

The temperature factor for C4 (argument c4 = TRUE) photosynthesis is calculated based on unpublished work as $$ \phi(T) = -0.008 + 0.00375 T - 0.58e-4 T^2 $$

The factor \(\phi(T)\) is to be multiplied with leaf absorptance and the fraction of absorbed light that reaches photosystem II. In the P-model these additional factors are lumped into a single apparent quantum yield efficiency parameter (argument kphio to function rpmodel).

References

Bernacchi, C. J., Pimentel, C., and Long, S. P.: In vivo temperature response func-tions of parameters required to model RuBP-limited photosynthesis, Plant Cell Environ., 26, 1419–1430, 2003

Examples

## Relative change in the quantum yield efficiency 
## between 5 and 25 degrees celsius (percent change):
print(paste((calc_ftemp_kphio(25.0)/calc_ftemp_kphio(5.0)-1)*100 ))

#> [1] "52.0396912899669"