Calculates the instantaneous temperature response of Jmax

Given Jmax at a reference temperature (argument tcref) this function calculates its temperature-scaling factor following modified Arrhenius kinetics based on Kattge & Knorr (2007). Calculates \(f\) for the conversion $$ V = f Vref $$

ftemp_inst_jmax(tcleaf, tcgrowth = tcleaf, tcref = 25)

Arguments

tcleaf

Leaf temperature, or in general the temperature relevant for photosynthesis (degrees Celsius)

tcgrowth

(Optional) Growth temperature, in the P-model, taken to be equal to tcleaf (in degrees Celsius). Defaults to tcgrowth = tcleaf.

tcref

Reference temperature (in degrees Celsius)

Value

A numeric value for \(fv\)

Details

The function is given by Kattge & Knorr (2007) as $$ fv = f(T, \Delta Hv) A/B $$ where \(f(T, \Delta Hv)\) is a regular Arrhenius-type temperature response function (see ftemp_arrh) with \(Hv=49884\) J mol-1, $$ A = 1 + exp( (T0 \Delta S - Hd) / (T0 R) ) $$ and $$ B = 1 + exp( (T \Delta S - Hd) / (TK R) ) $$ Here, \(T\) is in Kelvin, \(T0=293.15\) K, \(Hd = 200000\) J mol-1 is the deactivation energy and \(R\) is the universal gas constant and is 8.3145 J mol-1 K-1, and $$ \Delta S = aS - bS T $$ with \(aS = 659.70\) J mol-1 K-1, and \(bS = 0.75\) J mol-1 K-2, and \(T\) given in degrees Celsius (!)

References

Kattge, J. and Knorr, W.: Temperature acclimation in a biochemical model of photosynthesis: a reanalysis of data from 36 species, Plant, Cell and Environment, 30,1176–1190, 2007.

Examples

# Relative change in Jmax going (instantaneously, i.e.
# not acclimatedly) from 10 to 25 degrees (percent change):
print((ftemp_inst_jmax(25)/ftemp_inst_jmax(10)-1)*100 )
#> [1] 162.7662