Instantaneous scaling of rates

Calculates instantaneous scaling of rates, given acclimated capacities returned by function rpmodel.

inst_rpmodel(
  x,
  tc,
  vpd,
  co2,
  fapar,
  ppfd,
  patm = NA,
  elv = NA,
  kphio,
  verbose = FALSE
)

Arguments

x
An object (list) returned by function rpmodel.

tc
Temperature, relevant for photosynthesis (deg C)

vpd
Vapour pressure deficit (Pa)

co2
Atmospheric CO2 concentration (ppm)

fapar
(Optional) Fraction of absorbed photosynthetically active radiation (unitless, defaults to NA)

ppfd
(Optional) Photosynthetic photon flux density (mol m-2 d-1, defaults to NA). Note that the units of ppfd (per area and per time) determine the units of outputs lue, gpp, vcmax, and rd. For example, if ppfd is provided in units of mol m-2 month-1, then respective output variables are returned as per unit months.

patm
Atmospheric pressure (Pa). When provided, overrides elv, otherwise patm is calculated using standard atmosphere (101325 Pa), corrected for elevation (argument elv), using the function patm.

elv
Elevation above sea-level (m.a.s.l.). Is used only for calculating atmospheric pressure (using standard atmosphere (101325 Pa), corrected for elevation (argument elv), using the function patm), if argument patm is not provided. If argument patm is provided, elv is overridden.

kphio
Apparent quantum yield efficiency (unitless). Defaults to 0.0817 for method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_soilmstress=FALSE, 0.0870 for method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_soilmstress=TRUE, and 0.0492 for method_jmaxlim="wang17", do_ftemp_kphio=FALSE, do_soilmstress=FALSE, corresponding to the empirically fitted value as presented in Stocker et al. (2019) Geosci. Model Dev. for model setup 'BRC', 'FULL', and 'ORG' respectively.

verbose
provide verbose feedback (default = FALSE)

Value

A list containing instantaneous rates calculated from acclimated photosyntetic capacities (Vcmax25,Jmax25, "average" gs). This includes :

vcmax

jmax

rd

Examples

if (FALSE) {
 dampen_vec(
  vec = 20 * (sin(doy*pi/(365)))^2 + rnorm(365, mean = 0, sd = 5),
  tau = 40 )
}