P-model

R implementation of the P-model and its corrolary predictions (Prentice et al., 2014; Han et al., 2017).

rpmodel(tc, vpd, co2, elv, kphio, beta = 146, fapar = NA, ppfd = NA,
  c4 = FALSE, method_optci = "prentice14", method_jmaxlim = "wang17",
  do_ftemp_kphio = TRUE, returnvar = NULL)

Arguments

tc
Temperature, relevant for photosynthesis (deg C)

vpd
Vapour pressure deficit (Pa)

co2
Atmospheric CO2 concentration (ppm)

elv
Elevation above sea-level (m.a.s.l.)

kphio
Quantum yield efficiency parameter

beta
Unit cost ratio. Defaults to 146.0.

fapar
(Optional) Fraction of absorbed photosynthetically active radiation (unitless, defaults to NA)

ppfd
(Optional) Photosynthetic photon flux density (mol/m2, defaults to NA)

c4
(Optional) A logical value specifying whether the C3 or C4 photosynthetic pathway is followed. Defaults to FALSE. If method_optci===TRUE, ci is assumed to be very large and lue = kphio * fapar * ppfd.

method_optci
(Optional) A character string specifying which method is to be used for calculating optimal ci:ca. Defaults to "prentice14". Available also "prentice14_num" for a numerical solution to the same optimization criterium as used for "prentice14".

method_jmaxlim
(Optional) A character string specifying which method is to be used for factoring in Jmax limitation. Defaults to "wang17", based on Wang Han et al. 2017 Nature Plants and (Smith 1937). Available is also "smith19", following the method by Smith et al., 2019 Ecology Letters, and "none" for ignoring effects of Jmax limitation.

do_ftemp_kphio
(Optional) A logical specifying whether temperature-dependence of quantum yield efficiency after Bernacchi et al., 2003 PCE is to be accounted for. Defaults to TRUE.

returnvar
(Optional) A character string of vector of character strings specifying which variables are to be returned (see return below).

Value

A named list of numeric values (including temperature and pressure dependent parameters of the photosynthesis model, P-model predictions, including all its corollary). This includes :

ca: Ambient CO2 expressed as partial pressure (Pa)

gammastar: Photorespiratory compensation point \(\Gamma*\), (Pa), see calc_gammastar.

kmm: Michaelis-Menten coefficient \(K\) for photosynthesis (Pa), see calc_kmm.

ns_star: Viscosity, normalised to its value at 25 degrees Celsius \(\eta* = \eta(T) / \eta(25 deg C)\) used to scale the unit cost of transpiration. Calculated following Huber et al. (2009).

chi: Optimal ratio of leaf internal to ambient CO2 (unitless). Derived following Prentice et al. (2014) as: $$ \chi = \Gamma* / ca + (1- \Gamma* / ca) \xi / (\xi + \sqrt D ) $$ with $$ \xi = \sqrt (\beta (K+ \Gamma*) / (1.6 \eta*)) $$ \(\beta\) is given by argument beta, \(K\) is kmm (see calc_kmm), \(\Gamma*\) is gammastar (see calc_gammastar). \(\eta*\) is ns_star. \(D\) is the vapour pressure deficit (argument vpd), \(ca\) is the ambient CO2 partial pressure (ca).

ci: Leaf-internal CO2 partial pressure (Pa), calculated as \((\chi ca)\).

lue: Light use efficiency (g C / mol photons), calculated as $$ LUE = \phi(T) \phi0 m' Mc $$ where \(\phi(T)\) is the temperature-dependent quantum yield efficiency modifier (calc_ftemp_kphio) if do_ftemp_kphio==TRUE, and 1 otherwise. \(phi0\) is given by argument kphio. \(m'=m\) if method_jmaxlim=="none", or $$ m' = m \sqrt( 1 - (c/m)^(2/3) ) $$ with \(c=0.41\) (Wang et al., 2017) if method_jmaxlim=="wang17". \(Mc\) is the molecular mass of CO2 (12.0107 g mol-1). \(m\) is given by $$ (ci - \Gamma*) / (ci + 2 \Gamma*) $$

gpp: Gross primary production (g C m-2), calculated as $$ GPP = Iabs LUE $$ where \(Iabs\) is given by fapar*ppfd (arguments), and is NA if fapar==NA or ppfd==NA. Note that gpp scales with absorbed light. Thus, its units depend on the units in which ppfd is given.

iwue: Intrinsic water use efficiency (iWUE, unitless), calulated as $$ iWUE = = ca (1-\chi)/1.6 $$

gs: Stomatal conductance (gs), calculated as $$ gs = A / (ca (1-\chi)) $$ where \(A\) is gpp\(/Mc\).

vcmax: Maximum carboxylation capacity \(Vcmax\) (mol CO2 m-2) at growth temperature (argument tc), calculated as $$ Vcmax = \phi(T) \phi0 Iabs n $$ where \(n\) is given by $$ n = (ci + K) / (ci + 2 \Gamma*) $$

vcmax25: Maximum carboxylation capacity \(Vcmax\) (mol CO2 m-2) normalised to 25 deg C following a modified Arrhenius equation, calculated as \(Vcmax25 = Vcmax / fv\), where \(fv\) is the instantaneous temperature response by Vcmax and is implemented by function calc_ftemp_inst_vcmax.

rd: Dark respiration \(Rd\) (mol CO2 m-2), calculated as $$ Rd = b0 Vcmax (fr / fv) $$ where \(b0\) is a constant and set to 0.015 (Atkin et al., 2015), \(fv\) is the instantaneous temperature response by Vcmax and is implemented by function calc_ftemp_inst_vcmax, and \(fr\) is the instantaneous temperature response of dark respiration following Heskel et al. (2016) and is implemented by function calc_ftemp_inst_rd.

actnv: Active metabolic leaf N (canopy-level: mol N/m2-ground)

References

Huber, M. L., Perkins, R. A., Laesecke, A., Friend, D. G., Sengers, J. V., Assael,M. J., Metaxa, I. N., Vogel, E., Mares, R., and Miyagawa, K.: New internationalformulation for the viscosity of H2O, Journal of Physical and Chemical ReferenceData, 38, 101–125, 2009

Prentice, I. C., Dong, N., Gleason, S. M., Maire, V., and Wright, I. J.: Balancingthe costs of carbon gain and water transport: testing a new theoretical frameworkfor plant functional ecology, Ecology Letters, 17, 82–91, 10.1111/ele.12211,http://dx.doi.org/10.1111/ele.12211, 2014.

Wang, H., Prentice, I. C., Keenan, T. F., Davis, T. W., Wright, I. J., Cornwell, W. K.,Evans, B. J., and Peng, C.: Towards a universal model for carbon dioxide uptake byplants, Nat Plants, 3, 734–741, 2017. Atkin, O. K., et al.: Global variability in leaf respiration in relation to climate, plant func-tional types and leaf traits, New Phytologist, 206, 614–636, doi:10.1111/nph.13253, https://nph.onlinelibrary.wiley.com/doi/abs/10.1111/nph.13253.

Heskel, M., O’Sullivan, O., Reich, P., Tjoelker, M., Weerasinghe, L., Penillard, A.,Egerton, J., Creek, D., Bloomfield, K., Xiang, J., Sinca, F., Stangl, Z., Martinez-De La Torre, A., Griffin, K., Huntingford, C., Hurry, V., Meir, P., Turnbull, M.,and Atkin, O.: Convergence in the temperature response of leaf respiration acrossbiomes and plant functional types, Proceedings of the National Academy of Sciences, 113, 3832–3837, doi:10.1073/pnas.1520282113,2016.

Examples

out_rpmodel <- rpmodel( tc=10, vpd=300, co2=300, elv=300, kphio=0.06 )