Calculates the CO2 compensation point

Calculates the photorespiratory CO2 compensation point in absence of dark respiration, \(\Gamma*\) (Farquhar, 1980).

calc_gammastar(tc, patm)

Arguments

tc
Temperature, relevant for photosynthesis (degrees Celsius)

patm
Atmospheric pressure (Pa)

Value

A numeric value for \(\Gamma*\) (in Pa)

Details

The temperature and pressure-dependent photorespiratory compensation point in absence of dark respiration \(\Gamma* (T,p)\) is calculated from its value at standard temperature (\(T0 = 25 \)deg C) and atmospheric pressure (\(p0 = 101325\) Pa), referred to as \(\Gamma*0\), quantified by Bernacchi et al. (2001) to 4.332 Pa (their value in molar concentration units is multiplied here with 101325 Pa to yield 4.332 Pa). \(\Gamma*0\) is modified by temperature following an Arrhenius-type temperature response function \(f(T, \Delta Ha)\) (implemented by calc_ftemp_arrh) with activation energy \(\Delta Ha = 37830\) J mol-1 and is corrected for atmospheric pressure \(p(z)\) (see calc_patm) at elevation \(z\). $$ \Gamma* = \Gamma*0 f(T, \Delta Ha) p(z) / p_0 $$ \(p(z)\) is given by argument patm.

References

Farquhar, G. D., von Caemmerer, S., and Berry, J. A.: A biochemical model of photosynthetic CO2 assimilation in leaves of C 3 species, Planta, 149, 78–90, 1980.

Bernacchi, C. J., Singsaas, E. L., Pimentel, C., Portis, A. R. J., and Long, S. P.:Improved temperature response functions for models of Rubisco-limited photosyn-thesis, Plant, Cell and Environment, 24, 253–259, 2001

Examples

print("CO2 compensation point at 20 degrees Celsius and standard atmosphere (in Pa):")

#> [1] "CO2 compensation point at 20 degrees Celsius and standard atmosphere (in Pa):"

print(calc_gammastar(20, 101325))

#> [1] 3.339251