Calculates an empirical soil moisture stress factor

Calculates an empirical soil moisture stress factor as a function of relative soil moisture (fraction of field capacity).

calc_soilmstress(soilm, meanalpha = 1, apar_soilm = 0, bpar_soilm = 0.685)

Arguments

soilm
Relative soil moisture as a fraction of field capacity (unitless). Defaults to 1.0 (no soil moisture stress).

meanalpha
Local annual mean ratio of actual over potential evapotranspiration, measure for average aridity. Defaults to 1.0.

apar_soilm
(Optional, used only if do_calc_soilmstress==TRUE) Parameter determining the sensitivity of the empirical soil moisture stress function. Defaults to 0.0, the empirically fitted value as presented in Stocker et al. (2019) Geosci. Model Dev. for model setup 'FULL' (corresponding to a setup with method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_calc_soilmstress=TRUE).

bpar_soilm
(Optional, used only if do_calc_soilmstress==TRUE) Parameter determining the sensitivity of the empirical soil moisture stress function. Defaults to 0.685, the empirically fitted value as presented in Stocker et al. (2019) Geosci. Model Dev. for model setup 'FULL' (corresponding to a setup with method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_calc_soilmstress=TRUE).

Value

A numeric value for \(\beta\)

Details

The soil moisture stress factor is calculated using a quadratic function that is 1 above soilm = 0.6 and has a sensitivity, given by the y-axis cutoff, (zero soil moisture), determined by average aridity (argument meanalpha) as: $$ \beta = q(\theta - \theta*)^2 + 1 $$ for \(\theta < \theta*\) and \(\beta = 1.0\) otherwise. \(\theta*\) is fixed at 0.6. \(q\) is the sensitivity parameter and is calculated as a linear function of average aridity, quantified by the local annual mean ratio of actual over potential evapotranspiration, termed \(\alpha\): $$ q=(\beta0-1)/(\theta*-\theta0)^2 $$ \(\theta0\) is 0.0, and $$ \beta0 = a + b \alpha $$ \(a\) is given by argument apar, \(b\) is given by argument bpar.

References

Stocker, B. et al. Geoscientific Model Development Discussions (in prep.)

Examples

## Relative reduction (%) in GPP due to soil moisture stress at 
## relative soil water content ('soilm') of 0.2:
print((calc_soilmstress(0.2)-1)*100 )

#> [1] -14