Invokes a P-model function call

R implementation of the P-model and its corrolary predictions (Prentice et al., 2014; Han et al., 2017).

rpmodel(
  tc,
  vpd,
  co2,
  fapar,
  ppfd,
  patm = NA,
  elv = NA,
  kphio = ifelse(do_ftemp_kphio, ifelse(do_soilmstress, 0.087182, 0.081785), 0.049977),
  beta = 146,
  soilm = 1,
  meanalpha = 1,
  apar_soilm = 0,
  bpar_soilm = 0.733,
  c4 = FALSE,
  method_optci = "prentice14",
  method_jmaxlim = "wang17",
  do_ftemp_kphio = TRUE,
  do_soilmstress = FALSE,
  returnvar = NULL,
  verbose = FALSE
)

Arguments

tc
Temperature, relevant for photosynthesis (deg C)

vpd
Vapour pressure deficit (Pa)

co2
Atmospheric CO2 concentration (ppm)

fapar
(Optional) Fraction of absorbed photosynthetically active radiation (unitless, defaults to NA)

ppfd
(Optional) Photosynthetic photon flux density (mol m-2 d-1, defaults to NA). Note that the units of ppfd (per area and per time) determine the units of outputs lue, gpp, vcmax, and rd. For example, if ppfd is provided in units of mol m-2 month-1, then respective output variables are returned as per unit months.

patm
Atmospheric pressure (Pa). When provided, overrides elv, otherwise patm is calculated using standard atmosphere (101325 Pa), corrected for elevation (argument elv), using the function calc_patm.

elv
Elevation above sea-level (m.a.s.l.). Is used only for calculating atmospheric pressure (using standard atmosphere (101325 Pa), corrected for elevation (argument elv), using the function calc_patm), if argument patm is not provided. If argument patm is provided, elv is overridden.

kphio
Apparent quantum yield efficiency (unitless). Defaults to 0.081785 for method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_soilmstress=FALSE, 0.087182 for method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_soilmstress=TRUE, and 0.049977 for method_jmaxlim="wang17", do_ftemp_kphio=FALSE, do_soilmstress=FALSE, corresponding to the empirically fitted value as presented in Stocker et al. (2019) Geosci. Model Dev. for model setup 'BRC', 'FULL', and 'ORG' respectively.

beta
Unit cost ratio. Defaults to 146.0 (see Stocker et al., 2019).

soilm
(Optional, used only if do_soilmstress==TRUE) Relative soil moisture as a fraction of field capacity (unitless). Defaults to 1.0 (no soil moisture stress). This information is used to calculate an empirical soil moisture stress factor (calc_soilmstress) whereby the sensitivity is determined by average aridity, defined by the local annual mean ratio of actual over potential evapotranspiration, supplied by argument meanalpha.

meanalpha
(Optional, used only if do_soilmstress==TRUE) Local annual mean ratio of actual over potential evapotranspiration, measure for average aridity. Defaults to 1.0.

apar_soilm
(Optional, used only if do_soilmstress==TRUE) Parameter determining the sensitivity of the empirical soil moisture stress function. Defaults to 0.0, the empirically fitted value as presented in Stocker et al. (2019) Geosci. Model Dev. for model setup 'FULL' (corresponding to a setup with method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_soilmstress=TRUE).

bpar_soilm
(Optional, used only if do_soilmstress==TRUE) Parameter determining the sensitivity of the empirical soil moisture stress function. Defaults to 0.685, the empirically fitted value as presented in Stocker et al. (2019) Geosci. Model Dev. for model setup 'FULL' (corresponding to a setup with method_jmaxlim="wang17", do_ftemp_kphio=TRUE, do_soilmstress=TRUE).

c4
(Optional) A logical value specifying whether the C3 or C4 photosynthetic pathway is followed. Defaults to FALSE. If TRUE, the leaf-internal CO2 concentration is assumed to be very large and \(m\) (returned variable mj) tends to 1, and \(m'\) tends to 0.669 (with c = 0.41).

method_optci
(Optional) A character string specifying which method is to be used for calculating optimal ci:ca. Defaults to "prentice14".

method_jmaxlim
(Optional) A character string specifying which method is to be used for factoring in Jmax limitation. Defaults to "wang17", based on Wang Han et al. 2017 Nature Plants and (Smith 1937). Available is also "smith19", following the method by Smith et al., 2019 Ecology Letters, and "none" for ignoring effects of Jmax limitation.

do_ftemp_kphio
(Optional) A logical specifying whether temperature-dependence of quantum yield efficiency after Bernacchi et al., 2003 is to be accounted for. Defaults to TRUE.

do_soilmstress
(Optional) A logical specifying whether an empirical soil moisture stress factor is to be applied to down-scale light use efficiency (and only light use efficiency). Defaults to FALSE.

returnvar
(Optional) A character string of vector of character strings specifying which variables are to be returned (see return below).

verbose
Logical, defines whether verbose messages are printed. Defaults to FALSE.

Value

A named list of numeric values (including temperature and pressure dependent parameters of the photosynthesis model, P-model predictions, including all its corollary). This includes :

ca: Ambient CO2 expressed as partial pressure (Pa)

gammastar: Photorespiratory compensation point \(\Gamma*\), (Pa), see calc_gammastar.

kmm: Michaelis-Menten coefficient \(K\) for photosynthesis (Pa), see calc_kmm.

ns_star: Change in the viscosity of water, relative to its value at 25 deg C (unitless). $$\eta* = \eta(T) / \eta(25 deg C)$$ This is used to scale the unit cost of transpiration. Calculated following Huber et al. (2009).

chi: Optimal ratio of leaf internal to ambient CO2 (unitless). Derived following Prentice et al. (2014) as: $$ \chi = \Gamma* / ca + (1- \Gamma* / ca) \xi / (\xi + \sqrt D ) $$ with $$ \xi = \sqrt (\beta (K+ \Gamma*) / (1.6 \eta*)) $$ \(\beta\) is given by argument beta, \(K\) is kmm (see calc_kmm), \(\Gamma*\) is gammastar (see calc_gammastar). \(\eta*\) is ns_star. \(D\) is the vapour pressure deficit (argument vpd), \(ca\) is the ambient CO2 partial pressure in Pa (ca).

ci: Leaf-internal CO2 partial pressure (Pa), calculated as \((\chi ca)\).

lue: Light use efficiency (g C / mol photons), calculated as $$ LUE = \phi(T) \phi0 m' Mc $$ where \(\phi(T)\) is the temperature-dependent quantum yield efficiency modifier (calc_ftemp_kphio) if do_ftemp_kphio==TRUE, and 1 otherwise. \(\phi 0\) is given by argument kphio. \(m'=m\) if method_jmaxlim=="none", otherwise $$ m' = m \sqrt( 1 - (c/m)^(2/3) ) $$ with \(c=0.41\) (Wang et al., 2017) if method_jmaxlim=="wang17". \(Mc\) is the molecular mass of C (12.0107 g mol-1). \(m\) is given returned variable mj. If do_soilmstress==TRUE, \(LUE\) is multiplied with a soil moisture stress factor, calculated with calc_soilmstress.

mj: Factor in the light-limited assimilation rate function, given by $$ m = (ci - \Gamma*) / (ci + 2 \Gamma*) $$ where \(\Gamma*\) is given by gammastar.

mc: Factor in the Rubisco-limited assimilation rate function, given by $$ mc = (ci - \Gamma*) / (ci + K) $$ where \(K\) is given by kmm.

gpp: Gross primary production (g C m-2), calculated as $$ GPP = Iabs LUE $$ where \(Iabs\) is given by fapar*ppfd (arguments), and is NA if fapar==NA or ppfd==NA. Note that gpp scales with absorbed light. Thus, its units depend on the units in which ppfd is given.

iwue: Intrinsic water use efficiency (iWUE, Pa), calculated as $$ iWUE = ca (1-\chi)/(1.6) $$

gs: Stomatal conductance (gs, in mol C m-2 Pa-1), calculated as $$ gs = A / (ca (1-\chi)) $$ where \(A\) is gpp\(/Mc\).

vcmax: Maximum carboxylation capacity \(Vcmax\) (mol C m-2) at growth temperature (argument tc), calculated as $$ Vcmax = \phi(T) \phi0 Iabs n $$ where \(n\) is given by \(n=m'/mc\).

vcmax25: Maximum carboxylation capacity \(Vcmax\) (mol C m-2) normalised to 25 deg C following a modified Arrhenius equation, calculated as \(Vcmax25 = Vcmax / fv\), where \(fv\) is the instantaneous temperature response by Vcmax and is implemented by function calc_ftemp_inst_vcmax.

jmax: The maximum rate of RuBP regeneration () at growth temperature (argument tc), calculated using $$ A_J = A_C $$

rd: Dark respiration \(Rd\) (mol C m-2), calculated as $$ Rd = b0 Vcmax (fr / fv) $$ where \(b0\) is a constant and set to 0.015 (Atkin et al., 2015), \(fv\) is the instantaneous temperature response by Vcmax and is implemented by function calc_ftemp_inst_vcmax, and \(fr\) is the instantaneous temperature response of dark respiration following Heskel et al. (2016) and is implemented by function calc_ftemp_inst_rd.

Additional variables are contained in the returned list if argument method_jmaxlim=="smith19"

omega: Term corresponding to \(\omega\), defined by Eq. 16 in Smith et al. (2019), and Eq. E19 in Stocker et al. (2019).

omega_star: Term corresponding to \(\omega^\ast\), defined by Eq. 18 in Smith et al. (2019), and Eq. E21 in Stocker et al. (2019).

References

Bernacchi, C. J., Pimentel, C., and Long, S. P.: In vivo temperature response func-tions of parameters required to model RuBP-limited photosynthesis, Plant Cell Environ., 26, 1419–1430, 2003

Heskel, M., O’Sullivan, O., Reich, P., Tjoelker, M., Weerasinghe, L., Penillard, A.,Egerton, J., Creek, D., Bloomfield, K., Xiang, J., Sinca, F., Stangl, Z., Martinez-De La Torre, A., Griffin, K., Huntingford, C., Hurry, V., Meir, P., Turnbull, M.,and Atkin, O.: Convergence in the temperature response of leaf respiration across biomes and plant functional types, Proceedings of the National Academy of Sciences, 113, 3832–3837, doi:10.1073/pnas.1520282113,2016.

Huber, M. L., Perkins, R. A., Laesecke, A., Friend, D. G., Sengers, J. V., Assael,M. J., Metaxa, I. N., Vogel, E., Mares, R., and Miyagawa, K.: New international formulation for the viscosity of H2O, Journal of Physical and Chemical ReferenceData, 38, 101–125, 2009

Prentice, I. C., Dong, N., Gleason, S. M., Maire, V., and Wright, I. J.: Balancing the costs of carbon gain and water transport: testing a new theoretical frameworkfor plant functional ecology, Ecology Letters, 17, 82–91, 10.1111/ele.12211,http://dx.doi.org/10.1111/ele.12211, 2014.

Wang, H., Prentice, I. C., Keenan, T. F., Davis, T. W., Wright, I. J., Cornwell, W. K.,Evans, B. J., and Peng, C.: Towards a universal model for carbon dioxide uptake by plants, Nat Plants, 3, 734–741, 2017. Atkin, O. K., et al.: Global variability in leaf respiration in relation to climate, plant func-tional types and leaf traits, New Phytologist, 206, 614–636, doi:10.1111/nph.13253, https://nph.onlinelibrary.wiley.com/doi/abs/10.1111/nph.13253.

Smith, N. G., Keenan, T. F., Colin Prentice, I. , Wang, H. , Wright, I. J., Niinemets, U. , Crous, K. Y., Domingues, T. F., Guerrieri, R. , Yoko Ishida, F. , Kattge, J. , Kruger, E. L., Maire, V. , Rogers, A. , Serbin, S. P., Tarvainen, L. , Togashi, H. F., Townsend, P. A., Wang, M. , Weerasinghe, L. K. and Zhou, S. (2019), Global photosynthetic capacity is optimized to the environment. Ecol Lett, 22: 506-517. doi:10.1111/ele.13210

Stocker, B. et al. Geoscientific Model Development Discussions (in prep.)

Examples

rpmodel( tc = 20, vpd = 1000, co2 = 400, fapar = 1, ppfd = 300, elv = 0)

#> $ca
#> [1] 40.53
#> 
#> $gammastar
#> [1] 3.339251
#> 
#> $kmm
#> [1] 46.09928
#> 
#> $ns_star
#> [1] 1.125361
#> 
#> $chi
#> [1] 0.694352
#> 
#> $mj
#> [1] 0.7123038
#> 
#> $mc
#> [1] 0.3340838
#> 
#> $ci
#> [1] 28.14209
#> 
#> $lue
#> [1] 0.2547515
#> 
#> $gpp
#> [1] 76.42545
#> 
#> $iwue
#> [1] 7.742446
#> 
#> $gs
#> [1] 0.513655
#> 
#> $vcmax
#> [1] 19.04646
#> 
#> $vcmax25
#> [1] 29.8967
#> 
#> $jmax
#> [1] 38.36589
#> 
#> $rd
#> [1] 0.3025656
#>